+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/application/exceptions/PersonAlreadyExistsException.java
@@ -1,0 +1,8 @@
+package com.nimio2024.nimio2024_tfg_polmorenocastelar.application.exceptions;
+
+public class PersonAlreadyExistsException extends Exception{
+
+        public PersonAlreadyExistsException(String message) {
+            super(message);
+        }
+}