+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/persistence/PersonRepository.java
@@ -2,8 +2,12 @@ package com.nimio2024.nimio2024_tfg_polmorenocastelar.persistence;
 
 import com.nimio2024.nimio2024_tfg_polmorenocastelar.domain.Person;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
 public interface PersonRepository extends JpaRepository<Person, Long> {
+
+    @Query("SELECT p FROM Person p WHERE p.person_DNI = ?1")
+    Person findByDni(String dni);
 }