+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/service/PersonService.java
@@ -20,6 +20,10 @@ public class PersonService {
         return personRepository.findById(personId).orElse(null);
     }
 
+    public Person getPersonByDni(String dni) {
+        return personRepository.findByDni(dni);
+    }
+
     public void deletePerson(Long personId) {
         personRepository.deleteById(personId);
     }